+++ Timelon/Timelon.App/App.xaml
@@ -1,4 +1,4 @@
-﻿<Application x:Class="TimelonWPF.App"
+﻿<Application x:Class="Timelon.App.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              StartupUri="MainWindow.xaml">
