+++ Timelon/Timelon.App/MainWindow.xaml
@@ -1,10 +1,10 @@
-﻿<Window x:Class="TimelonWPF.MainWindow"
+﻿<Window x:Class="Timelon.App.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:i="http://schemas.microsoft.com/xaml/behaviors"
-        xmlns:local ="clr-namespace:TimelonWPF"
+        xmlns:local ="clr-namespace:Timelon"
         mc:Ignorable="d"
         Title="MainWindow" Height="768" Width="1024"
         MinWidth="1024" MinHeight="768"
