+++ Timelon/Timelon.App/Theme/TextBlockTemplateTheme.xaml
