+++ Timelon/Timelon/Manager.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Xml.Serialization;
-using TimelonCl.Data;
+using Timelon.Data;
 
-namespace TimelonCl
+namespace Timelon
 {
     /// <summary>
     /// Менеджер списков карт и провайдер данных
